# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PostController API",
        "version": "1.0"
    },
    "paths": {
        "/api/": {},
        "/api/auth/login/": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "User login and get token.",
                "operationId": "068f12b93c65e63717e80c4e37a565fe",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "user@user.user"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0L2FwaS9hdXRoL2xvZ2luIiwiaWF0IjoxNjk5OTY1OTc5LCJleHAiOjE2OTk5Njk1NzksIm5iZiI6MTY5OTk2NTk3OSwianRpIjoiMWpVeVpGVWYzdlpZRnJDcyIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.dlAUe2hCZSviHBlkUlTw_R8YGyhUUZx-cSyyfNey2Mc"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Display a listing of the post resource.",
                "operationId": "3efb096f87f8f5121f5f3a09f36fb420",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Some Title"
                                                    },
                                                    "likes": {
                                                        "type": "string",
                                                        "example": 321
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/multi": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Store a newly created post resource in storage",
                "operationId": "715eeccdd5d5a58c21dd61891223e903",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "likes": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": [
                                {
                                    "title": "Some Title",
                                    "likes": 321
                                },
                                {
                                    "title": "Other Title",
                                    "likes": 2222
                                }
                            ]
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "first_item": {
                                                    "properties": {
                                                        "1": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "title": {
                                                                        "type": "string",
                                                                        "example": "Some Title"
                                                                    },
                                                                    "likes": {
                                                                        "type": "string",
                                                                        "example": 321
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "2": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "title": {
                                                                        "type": "string",
                                                                        "example": "Some Title"
                                                                    },
                                                                    "likes": {
                                                                        "type": "string",
                                                                        "example": 321
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "second_item": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "first_item": {
                                            "1": [],
                                            "2": []
                                        },
                                        "second_item": []
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create post resource.",
                "operationId": "68eb607e387f7c5ea76af2596d29f700",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some Title"
                                            },
                                            "likes": {
                                                "type": "integer",
                                                "example": 321
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some Title"
                                                },
                                                "likes": {
                                                    "type": "string",
                                                    "example": 321
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Display the specified post resource.",
                "operationId": "95db749965aa33796e50f5b755d44771",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "id of the post",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some Title"
                                                },
                                                "likes": {
                                                    "type": "string",
                                                    "example": 321
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Remove the specified post resource from storage.",
                "operationId": "129dfa0fad8cde205dad3e860f0c4193",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "id of the post",
                        "required": true,
                        "example": 3
                    }
                ],
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Delete"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Post"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "26581515fe57d6dd19d7eab61e15cbca",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "id of the post",
                        "required": true,
                        "example": 666
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some Title"
                                            },
                                            "likes": {
                                                "type": "integer",
                                                "example": 321
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some Title"
                                                },
                                                "likes": {
                                                    "type": "string",
                                                    "example": 321
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}